Move unified i/o classes to connect.py
